app/requirements.txt
Django>=2.2.0,<3.0
djangorestframework>=3.9.0,<3.10.0
psycopg2-binary==2.8.3
django-cors-headers>=3.2.0,<3.3.0
djangorestframework-jwt==1.11.0
requests==2.22.0
